fix 'dot' test helper failing to encode string input
When asking the `dot` helper function to process a graph from inline input to a
binary output format, the `universal_newlines=True` parameter would not be
passed to the underlying `subprocess.check_output` call, resulting in failure to
encode the input. That is, something like the following:

  dot("png:gd", source="digraph { a -> b; }")

would result in:

  self = <subprocess.Popen object at 0x7fbb862ae390>
      def _communicate(self, input, endtime, orig_timeout):
  …
  >           input_view = memoryview(self._input)
  E           TypeError: memoryview: a bytes-like object is required, not 'str'
  /usr/lib64/python3.6/subprocess.py:1519: TypeError

Note that this problem was latent, as no usage fitting the above scenario
currently exists. Something like this will be introduced in a future commit.

Gitlab: related to #1786