lib/common: [nfc] replace 'nToName' with 'itos'
This code was attempting to avoid dynamic string construction with low numbered
IDs. This is unnecessary when using a modern compiler that understands functions
like `snprintf` as built-ins and can achieve better optimization without this
trick.

Apart from de-duplicating code, this has the effect of making this code thread
safe where it was not before.