CI: remove the use of '$OS_ID' during testing
This commit alters the CI tests to now pull OS information from /etc/os-release
instead of from an ad hoc environment variable. This makes test case execution
independent from which/whether environment variables have been exported. The
intent with this is to reduce the likelihood of an issue like that fixed in
50db0c5e8da293d816d8b5c4620f2fa2d311d37a being reintroduced.