CI: remove metrics collection in Pylint job
The Pylint job was running two checks:

  1. Everything, with warning count piped to metrics.txt
  2. Everything except the check for FIXMEs, hard failing if there is any
     warning

The latter passes. That is, all Graphviz Python code is warning-free with the
exception of FIXMEs. Meanwhile FIXMEs are frequently _intentionally_ introduced
in test code, indicating a known open bug for which a test case is being added.
So metric collection in step 1 above is not relevant; we expect this number
often to go up. This commit removes the collection of this unused metric to
decrease CI complexity.