fix: teach Smyrna '-?' on macOS and Windows
This is the equivalent of cd80b6262c7e1b95307756f13c89a9f6b2900520 and friends
but for Smyrna. It was not previously detected that this problem also affects
Smyrna because it is not currently built on macOS or Windows in CI. This will be
improved in future.

Gitlab: #1852