remove 'WIN32_LEAN_AND_MEAN' manipulation
This macro is a legacy toggle to suppress parts of the Windows API surface in
order to accelerate compilation. In a parallel compilation of a code base the
size of Graphviz, it has a negligible effect. As always, Raymond Chen gives the
back story.¹ This commit removes these tricks to ease future maintenance.

¹ https://devblogs.microsoft.com/oldnewthing/20091130-00/?p=15863