gvpr: initialize a variable
Squash
warning: ‘alpha’ may be used uninitialized in this function

One branch of conditional statement leaves variable alpha
uninitialized. Actually uninitialized value is not used,
but compiler still complains. So just add one assignment to
assumed 0.

BTW, the code can be refactored and shortened.
But it is preferable to just minimize the changes.