xlib plugin handle_file_events: more tightly scope a long-lived dynamic buffer
This buffer was being retained in a long-lived `static` pointer, increasing peak
memory usage and making it harder to use tools like Valgrind and ASan with
Graphviz. This change makes it more obvious to both users and tools what this
buffer is used for.

Note this also tweaks an error message to be more informative.