Visio plugin: fix: remove typoed #pragma dependencies
It seems to have been an accidental omission that these were not removed in
2d95aab626184f35f779bbd02a000a992826047a when migrating to describing link
dependencies in the build system files. However, these dependencies are typoed
too (should be `cgraph.lib` not `graph.lib`), so it is unclear how they could
have ever worked. This seems to have gone undiscovered because the Visio plugin
is not integrated into the MS Build files nor built on any Windows platform in
CI. This changes in an upcoming commit, exposing:

  LINK : fatal error LNK1104: cannot open file 'graph.lib'

Gitlab: #1836