sparse: fix: avoid a use of 'itos'
Usage of `itos` in this way relies on lifetime extension of a struct member in
an rvalue. While these semantics exist in C11 and C++, they do not in C99. As a
result, this causes undefined behavior.

This commit effectively reverts 10638c445d1f12bb3fcd0c45946ec8a2e3af60af.

Gitlab: #2229