unify tests/ and rtest/
For a long time, there has been confusion in the source tree with tests/ and
rtest/ each containing part of the test suite. These were initially distinct,
but over time files in tests/ have begun importing or referring to sources in
rtest/. This has been a common point of confusion for newcomers to the project:
tests/ is the logical place the test suite should live, so they run that, but
then discover it is not the full test suite.

The diff for this commit is large, so it is challenging to see that it is not a
complex change:

  1. Merge rtest/Makefile.am into tests/Makefile.am

  2. `git mv rtest/* tests/*`

  3. Adjust all Pytest invocations from `… ci/tests.py rtest tests` to
     `… ci/tests.py tests`

  4. Adjust all other references to rtest/ to tests/ (e.g. .gitignore)

Now seems to be a brief period of calm when edits to the test suite are not
in-flight, so we have less chance of a painful merge conflict. The problem of a
large part of the rtest/rtest.py suite being disabled remains and is not
affected by this commit.