graphml2gv: undo micro-optimization in 'endElementHandler'
This code was deciding between using a static stack buffer and a dynamic heap
buffer depending on the length of the string it needed to deal with. This kind
of optimization is unnecessary on modern platforms with performant allocators.
The branching and structure duplication involved in this is often now a
_de-optimization_ instead of the intended optimization.

This commit undoes this trick and instead unconditionally uses heap allocation.