gxl2gv: avoid large on-stack buffer
MSVC warns:

  C:\GitLab-Runner\builds\graphviz\graphviz\cmd\tools\gxl2gv.c(694): warning
    C6262: Function uses '20020' bytes of stack:  exceeds /analyze:stacksize
    '16384'.  Consider moving some data to heap.
    [C:\GitLab-Runner\builds\graphviz\graphviz\cmd\tools\gxl2gv.vcxproj]

So this function attempts to allocate ~20KB of stack data. Most of this is due
to the `buf` array. Using this much stack can cause crashes in constrained
environments. The buffer in question is used to read XML data in chunks, so we
can instead use the C standard library’s constant `BUFSIZ` that tells us the
optimal chunk size with which to perform I/O.

This change avoids the described problem as well as being a performance
optimization.